+++ include/zephyr/drivers/i2s.h
@@ -357,16 +357,6 @@ __subsystem struct i2s_driver_api {
 __syscall int i2s_configure(const struct device *dev, enum i2s_dir dir,
 			    const struct i2s_config *cfg);
 
-static inline int z_impl_i2s_configure(const struct device *dev,
-				       enum i2s_dir dir,
-				       const struct i2s_config *cfg)
-{
-	const struct i2s_driver_api *api =
-		(const struct i2s_driver_api *)dev->api;
-
-	return api->configure(dev, dir, cfg);
-}
-
 /**
  * @brief Fetch configuration information of a host I2S controller
  *
@@ -526,16 +516,6 @@ __syscall int i2s_buf_write(const struct device *dev, void *buf, size_t size);
 __syscall int i2s_trigger(const struct device *dev, enum i2s_dir dir,
 			  enum i2s_trigger_cmd cmd);
 
-static inline int z_impl_i2s_trigger(const struct device *dev,
-				     enum i2s_dir dir,
-				     enum i2s_trigger_cmd cmd)
-{
-	const struct i2s_driver_api *api =
-		(const struct i2s_driver_api *)dev->api;
-
-	return api->trigger(dev, dir, cmd);
-}
-
 /**
  * @}
  */
@@ -544,6 +524,7 @@ static inline int z_impl_i2s_trigger(const struct device *dev,
 }
 #endif
 
+#include <zephyr/drivers/i2s/internal/i2s_impl.h>
 #include <zephyr/syscalls/i2s.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_I2S_H_ */
